improve error messages for server snafus
The main goal of this patch is to improve error messages when the server
encounters a 500 or doesn't return any JSON data in an expected format.

To do this I also refactored the _get, _put, _post methods into a single
_request method that takes a `method` argument, since the code for them
was almost exactly identical. This did require me to change a whole
bunch of tests, unfortunately.