make PluvoGenerator more versatile
Instead of just being an iterable object, PluvoGenerator is now much
more like a full-featured sequence. It supports slicing and indexing
over the result set, fetching pages from pluvo as needed. I've renamed
the class PluvoResultSet.

The motivation for this feature is to allow Pluvo result sets to be easily
and efficiently used in e.g. the django Paginator class.

Note that offset and limit parameters are no longer supported (they are
ignored completely). The same effect can be achieved with very little
efficiency loss by slicing the result set instead, but this does
(probably?) constitute a backwards incompatible change.