turn PluvoResultSet into new-style class for python2.x
There is a small but significant difference between old and new style
classes when it comes to slicing: When omitting the start and stop
values in a slice, like so:

    the_list[:]

__getitem__ will be called with a slice item. In new style classes
slice.start and slice.stop wil both be equal to None. However, in old
style classes slice.start will equal 0, and slice.stop will equal
sys.maxint(!!). This, combined with the fact that we call range(start,
stop) in the __getitem__ method (which allocates a list in python2)
lead to us allocating about sys.maxint integers, causing a memory error.